Failed Interatcion
Ok so I looked up a turtorial and the script isn't working. I also realize that this would work later on in the project if we have an inventory or objects besides the terminal. I'm pushing anyway just so I can get some feed back, also don't spend to much time on it unless you can fix it.